requirements.txt
alembic==1.0.10
aniso8601==9.0.0
appdirs==1.4.4
atomicwrites==1.4.0
attrs==20.3.0
bcrypt==3.2.0
black==19.3b0
blinker==1.4
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==7.1.2
colorclass==2.2.0
coverage==5.4
cryptography==3.4.6
docopt==0.6.2
dparse==0.5.1
entrypoints==0.3
flake8==3.7.7
flasgger==0.9.2
Flask==1.0.3
Flask-Bcrypt==0.7.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
idna==2.8
importlib-metadata==3.4.0
isort==4.3.20
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.2.0
Mako==1.1.4
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.7.0
numpy==1.20.1
packaging==20.9
pip-upgrader==1.4.12
pluggy==0.13.1
psycopg2==2.8.2
py==1.10.0
pycodestyle==2.5.0
pycparser==2.20
pyflakes==2.1.1
PyJWT==2.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==4.6.2
pytest-cov==2.7.1
pytest-sugar==0.9.2
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.1
PyYAML==5.4.1
requests==2.22.0
safety==1.8.5
setproctitle==1.1.10
six==1.12.0
SQLAlchemy==1.3.4
termcolor==1.1.0
terminaltables==3.1.0
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.25.11
wcwidth==0.2.5
Werkzeug==0.15.4
zipp==3.4.0
